Extract helper methods into shared scope
