+++ EssentialFeed/EssentialFeedTests/Feed Api/URLSessionHTTPClientTests.swift
@@ -131,16 +131,8 @@ class URLSessionHTTPClientTests: XCTestCase {
         return receivedResult
     }
     
-    private func anyURL() -> URL {
-        return URL(string: "https://any-url.com")!
-    }
-    
     private func anyData() -> Data {
         return Data("any data".utf8)
-    }
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "any error", code: 0)
     }
     
     private func nonHTTPURLResponse () -> URLResponse {
